+++ tests/test_fldmean.py
@@ -38,7 +38,6 @@ class TestFldmean():
         #assert avg[1] == pytest.approx(17.9806)
         assert avg[1] == pytest.approx(300.1865)
 
-
     def test_fldmean_healpix(self):
         """Fldmean test for FESOM"""
         reader = Reader(model="ICON", exp="test-healpix", source='short')
@@ -46,6 +45,15 @@ class TestFldmean():
         avg = reader.fldmean(data['2t']).values
         assert avg.shape == (2,)
         assert avg[1] == pytest.approx(286.1479)
+
+    def test_fldmean_healpix_selection(self):
+        """Fldmean test for FESOM"""
+        reader = Reader(model="ICON", exp="test-healpix", source='short', regrid='r200')
+        data = reader.retrieve()
+        data = reader.regrid(data)
+        avg = reader.fldmean(data['2t'],  lon_limits=[-30,50], lat_limits=[-30, -90]).values
+        assert avg.shape == (2,)
+        assert avg[0] == pytest.approx(285.131484)
 
 
     def test_fldmean_icon(self):